requirements/prod.txt
# Everything needed in production

# Algorithms
torch==0.4.*

# Additional Information
py-cpuinfo==4.*

# Communication
pyzmq==17.*
aiozmq==0.8.*
json-rpc==1.11.*

# Data
pandas==0.23.*
simplejson==3.*

# Health Check
lorem==0.1.1